IOS-2360 Refactoring BlockBook API service
